Improve constants config
+ Bump dep
